// api/models/FeedbackCommentResponseData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FeedbackCommentResponseData = {
    id: number;
    feedback_id: number;
    gambler_id: number;
    author_name: string;
    comment: string;
    created_at: string;
    updated_at: string;
    viewer_is_author: boolean;
};