// api/models/VoteRequestData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * The direction tapped. Sending the same one again takes the vote back.
 */
export type VoteRequestData = {
    value: VoteRequestData.value;
};
export namespace VoteRequestData {
    export enum value {
        '_-1' = -1,
        '_1' = 1,
    }
}